разбивает код из form на модули
