+++ js/main.js
@@ -3,6 +3,7 @@ import {renderThumbnails} from './thumbnail-rendering.js';
 import './full-screen-picture.js';
 import './form.js';
 import './filters.js';
+import './scale.js';
 
 const similarPictures = createPictures();
 renderThumbnails(similarPictures);
